Move helper functions for editor to extensions
Moved GetLine, SetCol, StringToRunes and ToRunes to Extensions

This is to be able to call the methods from  EditorContext
